Add functions to print sidebar content
A loader for html was added to import html as text to generate folders
and todo items in sidebar. Events for #new-folder-button were added to
create new folders. Modified storage.js to return an error when a
invalid folder name is passed to addFolder method.

A todoSamples script created to create some todos and a folder for
testing.