Print todo information
Added a event for todo items to print its information into #todo
element and more events for the buttons inside of it.

Modified some function to no longer depend on the event object and
changed some names of functions and variables for more clearance. The
handleSubmission function was modified to also handle information
modification of an already existing todo, and openModal was modified
to place information of a todo into modalForm fields if the user wants
to edit the todo.

editTodo method from storage.js was mofidied because by mistake it set
finished property to false when finished argument was not present.